+++ components/core/include/memfault/core/reboot_reason_types.h
@@ -15,20 +15,31 @@ extern "C" {
 typedef enum MfltResetReason {
   kMfltRebootReason_Unknown = 0x0000,
 
+  //
   // Normal Resets
+  //
+
   kMfltRebootReason_UserShutdown = 0x0001,
   kMfltRebootReason_UserReset = 0x0002,
   kMfltRebootReason_FirmwareUpdate = 0x0003,
   kMfltRebootReason_LowPower = 0x0004,
   kMfltRebootReason_DebuggerHalted = 0x0005,
   kMfltRebootReason_ButtonReset = 0x0006,
+  kMfltRebootReason_PowerOnReset = 0x0007,
 
+  //
   // Error Resets
+  //
+
+  // Can be used to flag an unexpected reset path. i.e NVIC_SystemReset()
+  // being called without any reboot logic getting invoked.
   kMfltRebootReason_UnknownError = 0x8000,
   kMfltRebootReason_Assert = 0x8001,
   kMfltRebootReason_Watchdog = 0x8002,
+  kMfltRebootReason_BrownOutReset = 0x8003,
+  kMfltRebootReason_Nmi = 0x8004, // Non-Maskable Interrupt
 
-  // Arm Faults
+  // Resets from Arm Faults
   kMfltRebootReason_BusFault = 0x9100,
   kMfltRebootReason_MemFault = 0x9200,
   kMfltRebootReason_UsageFault = 0x9300,